(PDB-4512) engine: attempt to drop unused joins when requested
Add an opt-in optimization pass that examines query plans and attempts
to drop any top-level joins that don't appear to be necessary, where
necessary currently means that the join is required by  either the
requested projections or the top-level where clauses.

Attempt to make the optimizer pessimistic, such that it will bail out
and do nothing as soon as it hits anything it doesn't expect in the
plan.

Currently, each query has to opt-in by setting :can-drop-unused-joins?
to true, which should only be done if it seems clear that dropping any
of that query's top-level joins should be OK.  See the engine comments
for additional information about what "OK" may mean.

Assuming a query has opted-in, the optimization can be requested by
specifying the optimize_drop_unused_joins query parameter to true.

Add an intentionally undocumented PDB_QUERY_OPTIMIZE_DROP_UNUSED_JOINS
environment variable, that when set to "always" will cause pdb to act
as if every query had specified optimize_drop_unused_joins=true.  Use
this to run all the tests in travis with the optimization enabled as a
confidence building measure.

Right now, all the tests should work, but it's possible we'll have to
make adjustments for future test queries if they need to do something
that isn't compatible with the optimization.

Perhaps worth also noting that this code currently just does a
case-by-case analysis/matching of the plan structures as they
currently are, and is consequently harder to follow than we might
like.  I suspect given some time, and assuming the optimization ends
up being desirable, we could rework the engine (and/or plan
structures/specifications) to make this a lot cleaner (or even deeper
changes might make the optimization unnecessary).