(PDB-4600) Update sql query used in migration 73
Prior to this commit it was possible for PDB to OOM if it didn't have a
sufficient heap to track unique hashes being inserted into the resource_events
table. These hashes were used to deduplicate resource_events during the
partitioning migration. These duplicates were caused in part by PUP-9586 and
addressed in PDB-4315. 

This commit removes the clojure deduplication code and does the
deduplication via the sql query used to gather resource_events during
the migration. These changes allowed PDB to complete the migration using the
default 256mb heap size. It's important to note that this query can cause
postgres to perform an external merge on disk which takes approx. the same
space as the current size of the resource_events table. Users are generally
encouraged to have roughly the size of their DB free on disk for things like
postgresql migrations so this shouldn't be a problem in most cases.