(PDB-4600) Update migration 73 to account for possible name column
Prior to this commit migration 73 wouldn't save any existing data in the
resource_events name column if it was present. The original version of
migration 69 added this column to resource_events before it was replaced
by migration 73. This means that the name column may or may not be present
during migration 73 depending on which version of migration 69 a user
had already applied. Migration 73 now checks for the existence of a name
column in resource_events and migrates any existing name data.