(PDB-4479) Omit "gc-able" nodes from queries for inactive nodes
This modifies the definition of inactive_nodes in the query engine to
omit nodes that are eligible for garbage collection. These nodes are in
essence 'dead', in that the only reason we haven't deleted them is we
haven't hit the user's gc-interval.

We add an explicit definition of active_nodes because an "active node"
was previosuly defined as the complement of inactive_nodes and we
obviously don't want nodes that have been inactive for `node-purge-ttl`
and about to be garbage collected to temporarily enter "active" status.
The new definition for active_nodes is logically equivalent to the
complement of the _old_ definition of inactive_nodes because the
criteria for being an active node has not changed.

This solves a problem in pe-puppetdb where a node would be garbage
collected on one PuppetDB, but garbage collection had not yet run on the
other PuppetDB so that node was stil 'inactive' so the next sync run
would recreate and the node. This caused a never-ending ping pong effect
where a deactivated node would never be deleted from both PuppetDBs.
By omitting nodes that are eligible for garbage collection" from query
results for inactive_nodes, multiple PuppetDB's will agree on what
nodes are inactive regardless of when they last ran garbage collection.

For this change to solve the sync issue, it assumes that the times of
your PuppetDB's are in-sync, they have the same `node-purge-ttl`
setting, and when routing one command to multiple PuppetDB's the
command's producer_timestamp is set to a single timestamp for all copies
of the command. In our case, this is handled by the PuppetDB terminus.