(PDB-4626) date_trunc uses different semantics
The query to determine which partitions to create in advance of the
migration uses the postgresql function date_trunc, which will work and
match the trigger and other code if you do not attempt to coerce to UTC
at the same time.

Tested by switching my local machine to CET (UTC+01:00) and running the
tests, to match the customer use-case.