(PDB-4641) get unique dates properly in UTC
Force the timezone to UTC for determining the unique dates of
resource_events.

PostgreSQL returns dates in the client's timezone. What we need to do
here is force the connection into UTC. We want the date of the event, in
UTC, so that we can create the partition for it. To do this, we have the
find_resource_events_unique_dates plpgsql function above. For the
duration of this migration transaction, dates will be displayed in UTC.
If the timezone is in the local timezone (say, EST, or CEST), you get
the date in that timezone with the time cut off. So,
2020-01-30T05:00:00Z, which truncates to 2020-01-30 using the pgsql date
functions. In reality, this is actually 2020-01-31 in UTC, which is the
date we need for creating the partition successfully.