(PDB-4640) Add optional catalog resources file trgm index
The index is optional because it would introduce a breaking change for
FOSS users not following the recommendation of installing `pg_trgm` and
unwilling/unable to install it from the postgresql-contrib pacakge.

This index optional for now, this commit deprecates running without
`pg_trgm` and puppetdb will require `pg_trgm` in a later release.