(PDB-4672) Add periodic check for migration mismatch
This commit adds an at-at task that checks the current migration number
in PG and shuts PDB down if it detects that the database has been
migrated to a version the local PDB doesn't expect. This guards against
a situation that could arise with PDB on compilers where the database
is migrated to a newer version before the local PDB has been upgraded.