(maint) Fix testutil that clears db
If clear-db-for-testing! is called on a migrated db, it will attempt to
delete the sha1_agg aggregate function via the drop-function! testutils
method. Because aggregates and function are deleted differently (ie.
"DROP AGGREGATE" instead of "DROP FUNCTION") this would cause the test
to fail with

java.sql.BatchUpdateException: Batch entry 0 DROP FUNCTION IF EXISTS sha1_agg(bytea) CASCADE was aborted: ERROR: "sha1_agg" is an aggregate function
  Hint: Use DROP AGGREGATE to drop aggregate functions.

This adds a DROP AGGREGATE section to the clear-db-for-testing!
function. It also changes the sql-current-connection-function-names to
exclude aggregates and adds a new version of that function to lookup
aggregate names.

In PostgreSQL 11 the `proisagg` boolean column is replaced with
`prokind` char column. For `proisagg`, `'t'` means it is an aggregate
and `'f'` means it is not. For `prokind`, `'f'` means that it is a
function and `'a'` means that it is an aggregate.