(PDB-2398) Add delete-reports ezbake subcommand
This script must be run as root and requires the system to have either
`unzip` or `jar` on the `PATH`. Both are likely in pacakges that have to
be installed on the system, but `unzip` is likely a lot less intrustive
than installing the whole JDK.

Also, this assumes that PuppetDB and Postgres run on the same machine,
that you use the default `postgres` user.

The SQL command will delete all reports and resource_events
from the PuppetDB database, but prevents cascading to the certnames
table (which would then cascade to catalogs and factsets).

It also adds test designed to catch a few different cases.
1. We change the db schema in such a way that delete_reports.sql
   doesn't restore an identical schema
2. We delete node, catalog, of fact data that we did not want to
3. We didn't actually delete all the report/resource_events data