(PDB-4678) Update PostgreSQL SSL documentation
The main change is we will no longer document that FOSS users can use
their Puppet CA cert for postgres SSL via a Java keystore. They should
use libPQFactory because this is what is used in PE so it is more
tested. The documentation for creating a Java keystore remains, but is
downgraded to being recommended only if your cert comes from another CA
and is not able to be verified by the default Java keystore.