(PDB-4637) ext/test: edit/use a copy of the pdbbox pdb.ini
When we need to edit the pdb.ini, just edit/use a tempdir copy of the
pdbbox pdb.ini so we don't preclude the possibility of running
multiple tests (or a given test multiple times) using a single pdbbox,
perhaps an existing one.

Although right now that won't work for some of the tests because, for
example, they modify "puppetdb" rather than using a temporary database
name.