(PDB-4637) Ensure all db objects are owned by normal write user
When migration coordination is requested, make sure to set the db role
to the normal write user (not the migrator) after ejecting all the
non-migrators from the database, but before running any migrations, so
that all of the database objects created will be owned by a single
role.  That way the normal write user will still be able to drop
partitions during routine gc operations, etc.  (Since only the owner
or someone that's been granted the owner's role can drop a database.)