(PDB-4675) Fix deadlock issue in partitioned tables
In partitioned tables, we attempted every insertion to CREATE TABLE IF
NOT EXISTS and then INSERT the data. This caused a issue where when
processing multiple insertions concurrently, we could deadlock between
the table creation in one command processor and the insertion in
another.

This instead switched the CREATE TABLE step to be opportunistic, and
will only happen when an attempt to store the data failed with a "table
undefined" error. If multiple reports are being processed at once, we
can still see some errors in the postgresql logs about the table already
existing, but that is an example of rollbacks properly rectifying the
error.

By moving partition creation to another transaction, the insertion of
reports is returned to its previous complexity and the deadlocks no
longer occur.