(maint) Write pe.conf file to pdbbox
This is nearly identical to the example.conf in the
pe-puppetdb-extensions repo, with host, port, and passwords interpolated
to work with the pgbox that is generated.

Adds a pe.conf file to pdbbox-init, which will work with
pe-puppetdb-extensions. It still relies on the SSL in ./test-resources/
because the SSL files generated by the pdbbox-init scipt are not
generated by Puppet and don't work quite right when running pe-puppetdb