(PDB-4676) Update reports-partitioning migration to guard against OOM errors
Prior to this commit the reports-partitioning migration could possibly OOM
if the 500 reports it was pulling in per batch were too large. This commit
turns the FetchSize used in JDBC down to 1 and eliminates batching in the
migration. All reports are now processed serially and only one should be
held in memory at a time.

This change was tested on n2 against the SLV load test dataset which has
~345,000 reports, with the largest report rows being ~45kb in postgres.
With the change the migration was able to complete using a 256MB heap where
it previously would OOM. There seems to be a slight drop in performance,
but it appears to be less than a 10% difference across multiple runs
against this dataset.