(PDB-4637) Don't cache db metadata
Stop caching db metadata.  Change db-metadata from a delay to a
function that retrieves the data every time.  This fixes a problem
with the upcoming changes to the
unsupported-database-triggers-shutdown test, where the db-metadata
redef (to test that we shutdown when the db version is wrong) wouldn't
work because the call-with forms always redef again, causing the new
delay to retrieve the real data, not the tests's override data.

Assuming we're careful not to call db-metadata too frequently (and
perhaps if we are, we should just cache the value locally), this
approach is simpler and should be no worse and possibly better than
what we had before, since now it's at least possible for us to notice
database version changes.  We might also want to consider doing
something about them (PDB-4685).