(PDB-4637) Include user and username in all test configs
Set both the database user and username in all test configs so that
there can be no confusion (given our defaulting in config.clj and jdbc
make-connection-pool).  If nothing else, this avoids a problem in some
of our tests where the test was overrode the :user via a merge,
when the config already had a :username, and hilarity ensued.

In the longer run (perhaps next major release), we're planning to halt
with a config error if both user and username are set and differ.