(maint) Create files before background redirection
Redirection does create the file while the command is running, and does
so even before any output is produced. But the `&` operator takes
precendence here and runs the entire command in the background,
including the redirections, so its possible for the files to not
exist when they are read below.