(PDB-4701) allow unique top-level ast
The top-level ast syntax
```
["from" "reports" ["extract" ["certname"]]]
```
is only allowed at the top level.

This creates a top-level function that will optionally process the
extract statement, passing the inner expr, or whole expr if no extract
existed, to the maybe-add-agent-report-filter functions.

The top-level from clause is omitted because that is always unwrapped in
the query engine to look up an initial query-rec before processing a
user query to sql.