(PDB-4732) services: add initial support for multiple write dbs
Start adding support for multiple write databases by creating a new
scf-write-dbs entry in the global context that contains a vector of
database pools gnerated by a traversal of all of the write databases
defined in the config.  Add a parallel scf-write-db-names vector.

Forbid actually having more than one write database for now, and set the
venerable scf-write-db to be the one element in scf-write-dbs, which
will allow all the rest of the code to continue to work.