(PDB-4656) threadpools: rework to add an unbounded-threadpool
Add an unbounded-threadpool that's very similar to the existing gated
threadpool, but without the gate.  Rename some of the existing
functions to gated variants and change the namepsace alias from gtp to
pool, now that we have two kinds.

Allow the core thread to timeout too (via .allowCoreThreadTimeout),
since the command processing overhead should be much higher than the
startup overhead on the off chance that we suddenly get a command
after one minute completely idle.

Implement the Executor interface so that we can use ours like any
other, e.g. (.execute pool #(...)).