(PDB-4687) Deprecate java 8, recommend java 11
Modifies jdk-unsupported-msg to return either an error message or a
warning message for deprecated/unknown java versions. It will return an
error message for anything less than JDK 8, which will log the error and
shutdown pdb. For Java 8, 9, and 10, it will log a deprecation warning
recommending Java 11 and continue. For Java 12 it will log a warning
message stating this is not a tested version, recommend 11, and
continue.