(PDB-4743) Finish removing slingshot
The remaining uses are fairly straightforward, so finish the removal.
Checked to see that the schema validators do throw an ExceptionInfo of
:type ::schema.core/error, that trapperkeeper throws cli errors via
error-map, and that schema.utils/error-val is an exception we can just
rethrow.  Also observed that json-response does throw an ExceptionInfo
with a :message (happens during the tests).

The other changes were though to be safe because the affected code was
all inside puppetdb and/or the relevant exceptions aren't currently
being caught.