(PDB-4748) Update pinned deps for 6.11
Don't move pantomime past 2.5 because it causes fips dependency
problems, during say "lein with-profile fips test".

Don't move io.forward/yaml forward for now because newer versions of
snakecase appear to be incompatible with jdk 8, and it's not critical,
since it's only used to write configs in the integration tests.