(PDB-4658) Adjust storage metrics for cmd broadcast
This commit adjusts the handling of storage metrics to allow the creation
of multiple sets of metrics when in cmd broadcast mode. In broadcast
mode a full set of storage metrics will be created and prefixed with the db's
name. For example a db named 'foo' would produce a set of storage metrics
following the 'foo.metric-name' pattern. Metrics in normal mode should
remain unchanged.