(PDB-4658) Adjust admin metrics for cmd broadcast
This commit adjusts the handling of admin metrics to allow the creation
of multiple sets of metrics when in cmd broadcast mode. In broadcast
mode a full set of admin metrics will be created and prefixed with the db's
name. For example a db named 'foo' would produce a set of admin metrics
following the 'foo.metric-name' pattern. Metrics in normal mode should
remain unchanged.