(PDB-4812) convert-to-schema: always throw an exception
Previously this could crash when we attempted to throw the map
returned by error-val rather than an exception.

That can be demonstrated, for example, by adding an optional key to
one of the *-config-in maps in config, but not the
corresponding *-config-out map, and then attempting to use the value.

The exact case that uncovered it was the addition of this (which
wasn't quite right anyway) to per-database-config-in:

  :min-required-version (s/maybe [s/Int])