(PDB-4810) Exit on fatal exceptions and avoid silent failures
If an at-at interspaced task throws, the periodic invocations will be
canceled.  Guard against that, and guard a number of other
"independent" operations that might throw to the top-level and be
lost (e.g. futures).  Try to ensure we always at least report those
exceptions, and for "fatal" exceptions (see the new utils functions),
issue a shutdown request, e.g. for VirtualMachineError.