(PDB-4796) Fix race with blocking cmd submission during broadcast
This commit fixes a race during blocking cmd submission where the
callback was called with the result of the first completed broadcast
submission only. This could cause the client to receive a 503 if
one of the backends was down even if the cmd was stored successfully
in another. Blocking submission when broadcasting now follows the
same rules as async submission where storage in at least one backend
is counted as a success.