(PDB-4655) threadpool-test: check log event message for thread name
Instead of checking the log event's thread name directly via
.getThreadName, look for the thread name in the message (generated via
logging-exception-handler).

We'd noticed that the name was sometimes "main" in travis, and suspect
the problem may be a timing issue with the code in logback
LoggingEvent's getThreadName, which appears to wait to grab/cache the
current thread's name the first time you ask for it.

Perhaps related: https://jira.qos.ch/browse/LOGBACK-1088