(TK-487) For now, guard everything against deferred shutdowns
Guard all of the relevant service init and start methods against
pending shutdown requests by checking the
ShutdownService's get-shutdown-reason method, and arrange for all
other service methods (except stop) to throw an informative exception
if they're invoked during a pending shutdown (when the service
contexts might be incomplete).

This is just a stopgap until we see how TK-487 is resolved, which
would obviate more or less all of these changes.