(PDB-4839) Remove schema public from partition lookup
We don't prefix tables with their schema in any other query with its
schema, so we just assume that the PuppetDB tables (if you have tables
of the same name in a different schema) show up first in your schema
`search_path`. Omitting the prefix means we'll look for partitions of
the first 'reports' or 'resource_events' tables that we find on that
path.