(PDB-4745) disable gc in config-expiration test
Two things acquire the `clean-lock`, periodic garbage collection and the
admin clean endpoint. Garbage collection uses `(.lock clean-lock)`,
which means it will block until the lock is available. The admin
endpoint uses `(when (.tryLock clean-lock) <clean>)` which means if the
lock is not available it will return `false` and skip the attempt to
perform the admin clean action.  Sometimes, when this test runs, gc can
be running in the background and cause the initial call to
`(cli-svc/clean pdb ["expire_nodes"])` to be skipped, which puts the db
into an incorrect state for the following test checks.