(PDB-4853) with-shutdown-request-on-fatal-error: always rethrow
As a practical matter, swallowing the exception prevented it from
being printed from with-monitored-execution, but of course it also
prevented anything higher up the call stack from having the option to
respond.

Add an oom-causes-shutdown external test to test this, and to make
sure we actually do log OutOfMemoryErrors.  To facilitate that,
allocate memory from the schema check task when
PDB_TEST_ALLOCATE_AT_LEAST_MB_AT_STARTUP is set, so that we can
trigger the failure.

Always log our fatal error related shutdown requests (at least the
summary information -- the full backtrace should be logged by the
wrappers after we rethrow).