(PDB-4834) Add query AST/SQL logging
This commit adds support for query logging controlled by the log-queries
puppetdb config option. When log-queries is set to true and the debug
log level is set PuppetDB will log the incoming AST and SQL that is
generated for all queries. This logging is done for both external queries
and those submitted via the TK PuppetDBServer service's query method.