(PDB-4895) Add certnames catalog_inputs_hash column
This column will be a hash of the certname and a sorted set of the
inputs. Because most of the time the catalog inputs will be unchanged,
we will check for a change in the hash before making any changes to the
catalog_inputs table. This should drastically reduce the write-load and
overall churn on that table.