(PDB-4903) Set a lock_timeout for the report/event partition drops
This may well be obviated by future work to allow the gc to kick out
other queries, but for now, just make sure we don't wait indefinitely
for the exclusive lock we need when dropping a partition.

While we're waiting, all other queries that touch the "parent" table
will be blocked, which may be able to quickly block all command
processing threads, queries, etc., for the duration.