(PDB-4908) Add db "lock statuses" to track partition drops
Add a db-lock-status atom for each write database, initialized to

  {:write-locking-reports 0
   :write-locking-resource-events 0}

Increment the values whenever gc is waiting on or holding the relevant
write lock, i.e. whenever we're attempting to drop a related
partiiton.

This is currently not precise, i.e. a zero value doesn't guarantee no
one is waiting on or holding the related lock(s), but a non-zero value
guarantees that at least one thread is blocking on or holding such a
lock (either for the whole table or a partition thereof).

This can be used to allow other other operations to "get out of the
way", i.e. so that they don't hold up the gc (which itself blocks
table access attempts that come after it starts waiting).