(PDB-5003) Add (certname,end_time) index to reports
Prior to this commit, we had a certname index on reports
and an end_time index on reports.  For queries that want
certname sorted by end_time they were scanning the
endtime index and removing rows that don't match the certname
causing the query to read way more data pages into memory than
is desirable.

After this commit, we remove the certname index and replace it
with a certname,end_time index.  This allows for finding a certname
in the index and also getting the reports for it in the end_time
order that the query desires.  This drastically reduces the number
of data pages that need to be read into memory.

Also, any queries that previously used the certname index can still
use the certname,endtime index to satisfy their needs.