(PDB-4948) Allow GC to cancel blocking queries
This commit spawns a query-killer thread when the env var
PDB_GC_QUERY_KILLER_TIMEOUT_MS is set (defaults to 20 mins).
The query-killer is only spawned during periodic GC where we
expect contention with concurrent queries against the tables
that partition GC needs AccessExclusiveLocks in order to
drop. The query-killer loops and waits for the GC pid to get
blocked waiting on the lock it needs. When this state is
detected the query-killer will issue a pg_cancel_backend
call against all the pids from the PDB user against the PDB
database with granted locks. This should clear the way for
GC to get the lock it needs on the partitioned tables. This
can be disabled by setting PDB_GC_QUERY_KILLER_TIMEOUT_MS
to 0.