(PDB-4948) Allow GC to cancel blocking queries
This commit spawns a query-bulldozer thread when the env var
PDB_GC_QUERY_BULLDOZER_TIMEOUT_MS is set (defaults to 5 mins).
This setting controls the creation of the bulldozer thread and
how long the GC thread will wait for the bulldozer thread to be
cleaned up before logging an error.

The query-bulldozer is only spawned during periodic GC where we
expect contention with concurrent queries against the tables
that partition GC needs AccessExclusiveLocks in order to
drop. The query-bulldozer loops and waits for the GC pid to get
blocked waiting on the lock it needs. When this state is
detected the query-bulldozer will issue a pg_cancel_backend
call against all the pids from the PDB user against the PDB
database which are blocking GC. This should clear the way for
GC to get the lock it needs on the partitioned tables. This
can be disabled by setting PDB_GC_QUERY_BULLDOZER_TIMEOUT_MS
to 0.