(PDB-4974) Wait for schedulers to shut down (and replace at-at)
Although we'd expected to be waiting for our schedulers (at-at pools)
to stop before returning from our start/stop methods, we weren't
because at-at runs the relevant ScheduledThreadPoolExecutor shutdown
function in a future and returns immediately.

Since our needs are fairly straightforward, and since the relevant
blocking function in at-at is private, just manage the
ScheduledThreadPoolExecutors ourselves.

If any shutdown attempt times out, decide that we're in an unknown,
invalid state, and issue a request to shut down the whole server,
under the assumption that in that situation, it's better to exit and
be restarted.

cf. PDB-4932