(PDB-5003) Ensure migration 74 doesn't change over time
The reports partition migration (number 74) was using the runtime
create-reports-partition function, which as we add indexes, has changed
from the original function. Since migrations shouldn't change over time,
we now use a different version of the create partition function that
does not have the indexes added by later migrations.

This also removes the boolean flag on the report id index, which is no
longer necessary for testing because that migration will _always_ add
the index and _never_ be a noop now.