(PDB-5092) Update error handling in query response streaming
Prior to this commit if an error other than a java.sql.SQLException was
thrown in the munge-fn or at any time during query execution the query-error
promise wouldn't be delivered. This would cause the thread which called
deprecated-produce-streaming-body to block on the deref of the query-error
promise indefinitely, which also hung the client waiting on a response.
This commit changes the behavior to catch Throwable and forwards any errors
from the thread created in http/stream-response back to the caller. This
should guard against the query-error promise not being delivered if
something unexpected happens.