(PDB-2488) Allow empty query string
On endpoints underneath `/pdb/query/v4` having an empty query string was
valid. This would get parsed as "PQL" because it does not contain a "["
(even though PQL is not valid at sub-endpoints). There are 0 pql queries
in the string, so the previous check `(= 1 ...)` would fail.

This fixes the following query (which is used by puppetserver acceptance
tests) that previously worked.
```
curl -X POST http://localhost:8080/pdb/query/v4/catalog-input-contents \
  -H 'Content-Type:application/json' \
  -d '{"query": "" }'
```