(maint) test: fix pipefail error in jdk-info
The `echo` command needs to print multiple lines, but the `head -1`
command only reads the first line before completing. Sometimes this
means that the head command will finish and close its pipe before the
echo prints all its lines to the pipe. Due to the `-o pipefail` setting
in this file, this transient issue will cause the script to fail with
the error

```
+ ext/bin/check-spec-env lint/openjdk11
ext/bin/jdk-info: line 38: echo: write error: Broken pipe
Error: Process completed with exit code 1.
```

This allows the pipefail for this command to fix the transient error.