(PDB-5026) Switch context in query when using from
If the options (limit,offset,order_by) are provided to a
subquery that uses "from", the context isn't switched to the
entity from the "from" syntax. This behaviour results in an error
when the subquery is made on the reports table.

Ensure that context is correctly set and switched when a subquery
with "from" form is used.