(PDB-5134) Re-instate profiling for catalog submission
In PDB-5107 we accidentally removed the profiling for catalog submission
while attempting to optimize the memory usage of commands. This restores
profiling for the catalog submission command, but the metric will no
longer include the time it takes to serialize the Ruby hash that
represents the catalog to a JSON string.