(maint) dev: configure a migrator user
pdbbox-init now configures the puppetdb_migrator user in postgres and
populates the dev config so that it will be used by default.

The owner of the postgres db changes from the puppetdb user to the
postgres admin user because the puppetdb_migrator user cannot revoke the
connect priviledges from the database owner.