(docs) Remove the release_notes folder
Release notes that were moved to the `release_notes` folder no longer
had working relative links because they were in a different location in
the file tree.

Rename the release notes file for `6.x` to `release_notes_6.markdown`.
This should help with merge-ups to `main`.  Each time we add release
notes to `release_notes_6.markdown` it will now merge-up to the same
file in `main`.  Currently we get conflicts on every merge-up that adds
release notes to `release_notes.markdown` in `6.x` because the file with
the same name contains a totally different set of release notes on the
`main` branch.