(PDB-5216) Support a query "origin"
Add support for an optional query "origin", which can be provided via
a query parameters, and log it along with the query id.

Change the formatting of the query id log entries to be a json map,
e.g.

  2021-08-16 11:14:37,723 INFO  [qtp237850705-35] [p.p.query-eng]
  PDBQuery:6dd2a796-a4a3-44be-afb6-2a84459367f6:
  {"ast":["from","catalogs",["=","certname","host-3"]],"origin":"puppet:somewhere"}

to avoid any ambiguity, or the need for restrictions or a bespoke
encoding, with respect to the origin string contents.