(WWM-11655) Unusable pages
We suspect the issue in this ticket is being caused by behavior related
to pages called index.html. I've renamed the pages to overview to see if
that fixes it. It's better to avoid naming pages "index" anyway, cos I've
had other issues related to this. Once this is live, I'll create some
redirects for these pages as well.