(PDB-5259) Avoid fact_paths to improve fact-contents performance
Change the core fact-contents query from a cross join between factsets
and fact_paths to a single select from factsets.  In manual testing of
the core query, this reduced the explain analyze buffer reads by
orders of magnitude.

Instead of using the serialized "foo#~bar" path representation in
fact_paths to handle equality or regular expression path matches, just
generate the paths as a text[] with a recursive cte.

Within the cte, generate both a text[] and a type signature string
for each path, where the signature has an "s" for strings and an "i"
for array indexes for each corresponding text[] element, e.g.

   path: {"processors", "models", 0, "..."}
   types: "ssis"

To support queries against these rows, add a new :path-array type to
the engine, along with new expression types that allow "=", "in",
and "~>" matches.

To avoid risking trouble with the current engine limitations with
respect to query parameters, just produce properly escaped sql literal
values for some of the operators.

Match regular expressions componentwise.  This allows straightforward
support for regex operators like ^ and $, which is not provided by the
:encoded-path type (which the fact-paths endpoint still uses).
Document the current differences.