(PDB-5338) Strip unrecognized keys from catalog resources
Previously, if the agent wanted to add a new catalog resource attribute,
current and past versions of PuppetDB would fail to store the new
catalog due to the unexpected field. This could cause tight coupling of
Puppet Agent and PuppetDB versions in a non-major release, which is
undesirable.

To enable the agent to add the resource kind field, as well as future
resource fields without breaking PuppetDB, we need to strip the unknown
attributes from a catalog's resources because we do not currently have a
column to store them in.

For new resource attributes added in the future, PuppetDB will log one
warning message per-startup for any keys it does not recognize.